// lib/api/generated/model/unitDetailDtoEpisodesItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { UnitDetailDtoEpisodesItemStatus } from './unitDetailDtoEpisodesItemStatus';
import type { UnitDetailDtoEpisodesItemUserEpisode } from './unitDetailDtoEpisodesItemUserEpisode';

export type UnitDetailDtoEpisodesItem = {
  id: number;
  title: string;
  koreanTitle: string;
  description: string;
  status: UnitDetailDtoEpisodesItemStatus;
  userEpisode: UnitDetailDtoEpisodesItemUserEpisode;
};