// lib/api/generated/model/unitDetailDtoEpisodesItemStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */

export type UnitDetailDtoEpisodesItemStatus = typeof UnitDetailDtoEpisodesItemStatus[keyof typeof UnitDetailDtoEpisodesItemStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UnitDetailDtoEpisodesItemStatus = {
  COMPLETED: 'COMPLETED',
  AVAILABLE: 'AVAILABLE',
  LOCKED: 'LOCKED',
} as const;